Reintroduce cell.go
Now that I'm happier with the raw record
abstraction, I think we can put the cell stuff
back into its own file. This just makes the code
slightly more readable to me.
